backend/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
numpy==1.24.3
pandas==2.0.3
holidays==0.25
scikit-learn==1.3.2
xgboost==2.0.0